Simplify test renderer helper functions
Each time the new or another subcomponent is being tested, new helper
function had to be introduced, increasing the code size and making
helpers more redundant.
To simplify it, and make helpers more general, they are now rendering
the element (as JSX.Element) that is given as function's parameter,
instead of having it hardcoded each for a subcomponent.